api/requirements.txt
aniso8601==9.0.1
azure-core==1.23.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-RESTful==0.3.9
idna==3.3
isodate==0.6.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
msrest==0.6.21
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
